+++ config/initializers/active_admin.rb
@@ -15,6 +15,7 @@ require 'aa_erica_keywords'
 require 'aa_comment_paper_trail_patch'
 require 'aa_views_pages_base'
 require 'aa_hideable_column_mixin'
+require 'patches/active_admin/resource'
 
 ActiveAdmin::ResourceDSL.send(:include, ActiveAdmin::HideableColumnMixin::DSL)
 ActiveAdmin::ResourceDSL.send(:include, ActiveAdmin::ViewerCartMixin::DSL)
@@ -296,7 +297,7 @@ ActiveAdmin.setup do |config|
         if: proc { !menu['versions'].children.empty? }
       )
     end
-    
+
     admin.build_menu :utility_navigation do |menu|
       admin.add_current_user_to_menu (menu)
       menu.add  id: "logout", priority: 20,
@@ -304,13 +305,13 @@ ActiveAdmin.setup do |config|
                 html_options: { method: :delete },
                 url: -> { render_or_call_method_or_proc_on self, active_admin_namespace.logout_link_path },
                 if:  -> { current_active_admin_user? && !session["impersonated_user_id"] }
-              
+
       menu.add  id: "stop_impersonating",
                 priority: 20,
                 label: -> { "Stop Impersonating" },
                 url: -> { "/admin/users/stop_impersonating" },
                 if:  -> { current_active_admin_user? && session["impersonated_user_id"] }
-               
+
     end
   end